put back logic to create VNFC edges
put back logic to create VNFC edges
Consolidated similiar sim scenrios into 1 and updated robots properties
Support macro scerios for vnfc to vfmodule relationship.

Issue-ID: SO-3361

+++ so-simulator/src/main/java/org/onap/so/simulator/actions/aai/ProcessVnfc.java
@@ -1,11 +1,9 @@
 package org.onap.so.simulator.actions.aai;
 
 import java.util.Optional;
+import org.onap.aai.domain.yang.VfModule;
 import org.onap.aai.domain.yang.Vnfc;
-import org.onap.aai.domain.yang.Vserver;
-import org.onap.aai.domain.yang.Vservers;
 import org.onap.aaiclient.client.aai.AAIResourcesClient;
-import org.onap.aaiclient.client.aai.entities.uri.AAIPluralResourceUri;
 import org.onap.aaiclient.client.aai.entities.uri.AAIResourceUri;
 import org.onap.aaiclient.client.aai.entities.uri.AAIUriFactory;
 import org.onap.aaiclient.client.generated.fluentbuilders.AAIFluentTypeBuilder;
@@ -50,40 +48,28 @@ public class ProcessVnfc extends AbstractTestAction {
                 vnfc.setModelVersionId("9e314c37-2258-4572-a399-c0dd7d5f1aec");
                 vnfc.setModelCustomizationId("2bd95cd4-d7ff-4af0-985d-2adea0339921");
 
+                AAIResourceUri vfModuleURI = AAIUriFactory.createResourceUri(AAIFluentTypeBuilder.network()
+                        .genericVnf(context.getVariable("vnfId")).vfModule(context.getVariable("vfModuleId")));
+
                 if (aaiResourceClient.exists(vnfcURI)) {
-                    logger.debug("cleaning up VNFC");
-                    aaiResourceClient.delete(vnfcURI);
+                    Optional<VfModule> vfModule = aaiResourceClient.get(vfModuleURI).asBean(VfModule.class);
+                    if (vfModule.get().getVfModuleName().contains("macro")) {
+                        String vnfcName = "ssc_server_1" + vfModule.get().getVfModuleName()
+                                .substring(vfModule.get().getVfModuleName().length() - 1);
+                        vnfc.setVnfcName(vnfcName);
+                        vnfcURI = AAIUriFactory.createResourceUri(AAIFluentTypeBuilder.network().vnfc(vnfcName));
+                    } else {
+                        logger.debug("cleaning up VNFC");
+                        aaiResourceClient.delete(vnfcURI);
+                    }
                 }
 
                 logger.debug("creating new VNFC");
                 aaiResourceClient.create(vnfcURI, vnfc);
 
-                AAIResourceUri vfModuleURI = AAIUriFactory.createResourceUri(AAIFluentTypeBuilder.network()
-                        .genericVnf(context.getVariable("vnfId")).vfModule(context.getVariable("vfModuleId")));
                 logger.debug("creating VNFC edge to vf module");
                 aaiResourceClient.connect(vfModuleURI, vnfcURI);
-            } else if (context.getVariable("requestAction").equals("CreateVfModuleInstance")
-                    && context.getVariable("serviceAction").equals("activate")) {
-                logger.debug("creating edge between vserver and vnfc");
-                AAIResourceUri vnfcURI =
-                        AAIUriFactory.createResourceUri(AAIFluentTypeBuilder.network().vnfc("ssc_server_1"));
-                AAIPluralResourceUri vserverPlural =
-                        AAIUriFactory
-                                .createResourceUri(AAIFluentTypeBuilder.cloudInfrastructure()
-                                        .cloudRegion(context.getVariable("cloudOwner"),
-                                                context.getVariable("cloudRegion"))
-                                        .tenant(context.getVariable("tenant")).vservers())
-                                .queryParam("vserver-name", "ssc_server_1");
-                Optional<Vserver> vserver = aaiResourceClient.getFirst(Vservers.class, Vserver.class, vserverPlural);
-                if (vserver.isPresent()) {
-                    AAIResourceUri vserverURI =
-                            AAIUriFactory.createResourceUri(AAIFluentTypeBuilder.cloudInfrastructure()
-                                    .cloudRegion(context.getVariable("cloudOwner"), context.getVariable("cloudRegion"))
-                                    .tenant(context.getVariable("tenant")).vserver(vserver.get().getVserverId()));
-                    aaiResourceClient.connect(vserverURI, vnfcURI);
-                }
             }
-
         } catch (Exception e) {
             logger.debug("Exception in ProcessVnfc.doExecute", e);
         }
